Eliminar esta completado

+++ resources/js/components/ListLocations.vue
@@ -41,6 +41,29 @@
             </div>
 
         </div><!--/row-->
+
+        <!-- Modal -->
+        <div class="modal" tabindex="-1" role="dialog">
+            <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title">Modal title</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    <div class="modal-body">
+                        <p>Modal body text goes here.</p>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-primary">Save changes</button>
+                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+
     </div>
 </template>
 
@@ -62,14 +85,17 @@
                 console.log('listando los locations');
             },
             onClickDelete(id) {
-                let uri = `api/location/delete/${id}`;
-                this.axios.delete(uri).then(response => {
-                    this.locations.splice(this.locations.indexOf(id), 1);
-                    this.$toast.success({
-                        title:'Lugar Eliminado',
-                        message:'El Lugar eliminado con exito!!'
+                if(confirm('Realmente quiere eliminar el lugar?')) {
+                    let uri = `api/location/delete/${id}`;
+                    this.axios.delete(uri).then(response => {
+                        this.locations.splice(this.locations.indexOf(id), 1);
+                        this.$toast.success({
+                            title:'Lugar Eliminado',
+                            message:'El Lugar eliminado con exito!!'
+                        });
                     });
-                });
+                }
+
             },
             onClickDetail() {
                 console.log('Detallar ...');
